backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-openai
supabase
pydantic
python-dotenv
httpx
